+++ stores/auth.ts
@@ -1,10 +1,10 @@
-import axios from "axios";
+import apiClient from "@/providers/apiClient";
 import { ApiUrls } from "@/consts/ApiUrls";
-
 export default function authStore() {
   return {
     async login() {
-      await axios.post(ApiUrls.POST_AUTH);
+      // Firebase Functions に FID を送信
+      await apiClient.post(ApiUrls.POST_AUTH);
     },
   };
 }
